fix: do not attempt to mask invalid IP addresses
If an IP is invalid, then it may not be considered PII data, so
do not attempt to mask it and just return the current value.
Attempting to mask it will cause an error for invalid address.

It returns the same value as returning `nil` may cause issues with
`NOT NULL` columns.